requirements.txt
mcp>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv
dedalus_labs
aiohttp>=3.8.0
